more progress towards a better import rewriting system
