+++ tests/test_transformers.py
@@ -1,18 +1,43 @@
 import textwrap
 
 import pytest
-from unpy._transform import transform_source
 from unpy._types import PythonVersion
+from unpy.transformers import transform_source
 
 
 def _src(source: str, /) -> str:
-    return textwrap.dedent(source).lstrip("\n")
+    out = textwrap.dedent(source).lstrip("\n")
+    if not out.endswith("\n"):
+        out += "\n"
+    return out
+
+
+@pytest.mark.parametrize("source", ["", "\n", "    \n        \n\n\n"])
+def test_whitespace(source: str) -> None:
+    assert transform_source(source) == source
+
+
+@pytest.mark.parametrize("source", ["# comment", '"""docstring"""'])
+def test_comments(source: str) -> None:
+    assert transform_source(source) == source
+
+
+@pytest.mark.parametrize(
+    "source",
+    [
+        "import sys\nprint(*sys.argv)\n",
+        "__version__: str = '3.14'\n",
+        "def concat(*args: str) -> str: ...\n",
+        "class C:\n    def f(self, /) -> None: ...\n",
+        "raise NotImplementedError\n",
+    ],
+)
+def test_already_compatible(source: str) -> None:
+    assert transform_source(source) == source
 
 
 def test_type_alias_simple():
-    pyi_in = _src("""
-    type AnyStr = str | bytes
-    """)
+    pyi_in = _src("type AnyStr = str | bytes")
     pyi_expect = _src("""
     from typing import TypeAlias
     AnyStr: TypeAlias = str | bytes
@@ -37,9 +62,7 @@ def test_type_alias_simple_typing_import():
 
 
 def test_type_alias_param():
-    pyi_in = _src("""
-    type Pair[T] = tuple[T, T]
-    """)
+    pyi_in = _src("type Pair[T] = tuple[T, T]")
     pyi_expect = _src("""
     from typing import TypeAlias, TypeVar
 
@@ -51,9 +74,7 @@ def test_type_alias_param():
 
 
 def test_type_alias_param_bound():
-    pyi_in = _src("""
-    type Shape2D[N: int] = tuple[N, N]
-    """)
+    pyi_in = _src("type Shape2D[N: int] = tuple[N, N]")
     pyi_expect = _src("""
     from typing import TypeAlias, TypeVar
 
@@ -83,9 +104,7 @@ def test_type_alias_param_constraints():
 
 
 def test_type_alias_param_default():
-    pyi_in = _src("""
-    type OneOrMany[T = object] = T | tuple[T, ...]
-    """)
+    pyi_in = _src("type OneOrMany[T = object] = T | tuple[T, ...]")
     pyi_expect = _src("""
     from typing import TypeAlias
     from typing_extensions import TypeVar
@@ -98,9 +117,7 @@ def test_type_alias_param_default():
 
 
 def test_type_alias_params_order_mismatch():
-    pyi_in = _src("""
-    type RPair[T1, T0] = tuple[T0, T1]
-    """)
+    pyi_in = _src("type RPair[T1, T0] = tuple[T0, T1]")
     pyi_expect = _src("""
     from typing import TypeVar
     from typing_extensions import TypeAliasType
@@ -139,9 +156,7 @@ def test_type_alias_dupe_clash():
 
 
 def test_generic_function():
-    pyi_in = _src("""
-    def spam[T](x: T) -> T: ...
-    """)
+    pyi_in = _src("def spam[T](x: T) -> T: ...")
     pyi_expect = _src("""
     from typing import TypeVar
 
@@ -153,9 +168,7 @@ def test_generic_function():
 
 
 def test_generic_function_bound():
-    pyi_in = _src("""
-    def f[Z: complex](z: Z) -> Z: ...
-    """)
+    pyi_in = _src("def f[Z: complex](z: Z) -> Z: ...")
     pyi_expect = _src("""
     from typing import TypeVar
 
@@ -167,9 +180,7 @@ def test_generic_function_bound():
 
 
 def test_generic_function_constraints():
-    pyi_in = _src("""
-    def f[Z: (int, float, complex)](z: Z) -> Z: ...
-    """)
+    pyi_in = _src("def f[Z: (int, float, complex)](z: Z) -> Z: ...")
     pyi_expect = _src("""
     from typing import TypeVar
 
@@ -181,9 +192,7 @@ def test_generic_function_constraints():
 
 
 def test_generic_function_default():
-    pyi_in = _src("""
-    def f[Z: complex = complex](z: Z = ...) -> Z: ...
-    """)
+    pyi_in = _src("def f[Z: complex = complex](z: Z = ...) -> Z: ...")
     pyi_expect = _src("""
     from typing_extensions import TypeVar
 
@@ -247,9 +256,7 @@ def test_generic_function_dupe_clash_type():
 
 
 def test_generic_class():
-    pyi_in = _src("""
-    class C[T_contra, T, T_co]: ...
-    """)
+    pyi_in = _src("class C[T_contra, T, T_co]: ...")
     pyi_expect = _src("""
     from typing import Generic
     from typing_extensions import TypeVar
