+++ unpy/main.py
@@ -9,8 +9,8 @@ from typing import Annotated, Final, TypeAlias
 
 import typer
 
-from ._transform import transform_source
 from ._types import PythonVersion
+from .transformers import transform_source
 
 __all__ = ("app",)
